+++ brake/Core/Src/battery_sensor.c
@@ -15,7 +15,7 @@ BatterySensor *new_battery_sensor(ADCSensor *_adc_sensor, uint8_t _rank) {
 // REQUIRES: ForceSensor is a force_sensor object
 // MODIFIES: nothing
 // EFFECTS: Returns the currently stored value battery sensor data.
-float get_battery_sensor_data(BatterySensor *battery_sensor) {
+uint8_t get_battery_sensor_data(BatterySensor *battery_sensor) {
     uint32_t raw_value = get_adc_sensor_value(battery_sensor->adc_sensor, battery_sensor->rank);
     float voltage = raw_value * 3.3f / 64.0;
     if (voltage > BATTERY_SENSOR_FIVE_BATTERY) {
